fix missing N/A display for empty field fix missing underline for current page fix getting wrong ETD and ETA for date in the past fix missing status after clicking Achieve on ShipmentBooking and ConfirmArrival fix showing wrong ETD and ETA on VesselDeparture fix only allow editing for items that have not go through progress check fix display issue for large number on item Quantity fix item Quantity range: 1 - 9 999 999 999 (10 digits)

+++ ADJ-Internship/BusinessService/Dtos/ShipmentBookingDtos.cs
@@ -62,7 +62,7 @@ namespace ADJ.BusinessService.Dtos
     //DropList of Boxed/Carton
     public string PackType { get; set; }
 
-    public float Quantity { get; set; }
+    public decimal Quantity { get; set; }
 
     public float Cartons { get; set; }
 
@@ -126,7 +126,7 @@ namespace ADJ.BusinessService.Dtos
     public string Vendor { get; set; }
 
     [Display(Name = "PO Quantity")]
-    public float Quantity { get; set; }
+    public decimal Quantity { get; set; }
 
     //equal Revise Quantity
     [Display(Name = "Booking Quantity")]
